convertimelo/functions/requirements.txt
firebase-functions
firebase-admin
pint
google-generativeai
flask
flask-cors